+++ WebApi/App_Start/FilterConfig.cs
@@ -6,17 +6,14 @@
     {
         public static void RegisterFilters(HttpFilterCollection filters)
         {
-            // TODO Can't be enabled because "/odata/$metadata" fails too (which should be anonymously accessible) and cannot be set 'AllowAnonymous'
-            // Instead both 'Base' controllers (Api & OData) use Authorize filter
-
-            // Authorize required by default
-            //filters.Add(new AuthorizeAttribute());
-
             // DataServiceVersion header
             filters.Add(new Filters.DataServiceVersionHeaderAttribute());
 
-            // ValidateModel filter
+            // ValidateModel
             filters.Add(new Filters.ValidateModelAttribute());
+
+            // Exception handler
+            filters.Add(new Filters.ExceptionHandlerAttribute());
         }
     }
 }